refactoring, class to handle channel-related events
