+++ src/utils/Router.ts
@@ -1,8 +1,10 @@
-import EventHandler from "../handlers/EventHandler";
+import EventHandler from "../handlers/events/app_mentions/Events.handler";
+import ChannelHandler from "../handlers/events/channel/ChannelEvents.handler";
 import logger from './Logger';
 import { Response, Request } from 'express';
 
 const ev = new EventHandler();
+const ch = new ChannelHandler();
 
 /**
  * Given a request, route it to the appropriate handler.
@@ -14,6 +16,10 @@ export default async function RouteRequest(requestType: string, req: Request, re
     switch (requestType) {
         case "app_mention":
             ev.appMention(req, res);
+            break;
+        
+        case "channel_left":
+            ch.handleChannelLeft(req, res);
             break;
     
         default:
